Правки 8.14
