+++ js/thumbnail-rendering.js
@@ -1,4 +1,5 @@
 import { photos } from './create-photo-and-description.js';
+import { openBigPicture } from './open-big-photo.js';
 
 const pictureTemplate = document.querySelector('#picture')
   .content
@@ -6,6 +7,16 @@ const pictureTemplate = document.querySelector('#picture')
 const pictures = document.querySelector('.pictures');
 const pictureFragment = document.createDocumentFragment();
 const similarPhotos = photos;
+const pictureContainer = document.querySelector('.pictures');
+
+pictureContainer.addEventListener('click', (evt) => {
+  const targetPicture = evt.target.closest('.picture');
+
+  if (targetPicture) {
+    evt.preventDefault();
+    openBigPicture(targetPicture.dataset.pictureId);
+  }
+});
 
 similarPhotos.forEach((photo) => {
   const { id, url, description, comments, likes } = photo;
@@ -20,4 +31,3 @@ similarPhotos.forEach((photo) => {
 });
 
 pictures.appendChild(pictureFragment);
-
